requirements.txt
transformers==4.33.0
torch==2.0.1
requests
beautifulsoup4
google-search-results==2.4.2
python-dotenv
pytest
pytest-mock